scripts/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
schedule>=1.2.0
pyyaml>=6.0.1
python-dateutil>=2.8.2